+++ src/Grafikus/Aszteroida.java
@@ -1,7 +1,6 @@
-package Proto;
+package Grafikus;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.StringJoiner;
