+++ src/Grafikus/Hajo.java
@@ -1,4 +1,4 @@
-package Proto;
+package Grafikus;
 
 /**
  *
