+++ src/Grafikus/Jatek.java
@@ -1,4 +1,4 @@
-package Proto;
+package Grafikus;
 
 
 import java.io.File;
