+++ src/Grafikus/Leptetheto.java
@@ -1,4 +1,4 @@
-package Proto;
+package Grafikus;
 /**
  * A jatekban lepesre kepes objekumok kozos interfesze. A lepest a Lepes() fuggveny meghivasa jelenti az adott osztalyon
  *
