+++ src/Grafikus/Main.java
@@ -1,5 +1,5 @@
 
-package Proto;
+package Grafikus;
 
 public class Main {
 	public static void main(String[] args) {
