+++ src/Grafikus/Teleportkapu.java
@@ -1,4 +1,4 @@
-package Proto;
+package Grafikus;
 
 import java.util.ArrayList;
 import java.util.List;
