+++ src/Grafikus/Vas.java
@@ -1,4 +1,4 @@
-package Proto;
+package Grafikus;
 
 /**A jatekban szereplo vas nyersanyag*/
 public class Vas extends Nyersanyag{
